same refactoring of BasicTests
split each test to have its own class and rule as with moves

It is slower, but less prone to the concurency issue shuting down tests
in CI
